UPDATE
In this update I was improved how my classes and methods were arranged. I kept my properties in their own class called Credentials and I kept the methods under MyUserDictionary. This project is not completed. I still need to encrypt and decrypt the password that is passed in from the user.